improved upload
