+++ client/src/containers/Roster/hooks/usePaginatedChildData.tsx
@@ -29,7 +29,7 @@ export const usePaginatedChildData = (query: RosterQueryParams) => {
 
     // Paginated api query (but only once we have organizationId param)
     return query.organization
-      ? `/children?${stringify({
+      ? `children?${stringify({
           organizationId: query.organization,
           skip: index * PAGE_SIZE,
           take: PAGE_SIZE,
