+++ src/controllers/enrollmentReports/index.ts
@@ -1,5 +1,5 @@
 export { parseUploadedTemplate, parseEnrollmentReportRow } from './parse';
-export { mapRows } from './map';
+export { mapRows } from './map/map';
 export * from './removeExisting';
 export * from './checkErrors';
-export * from './mapUtils';
+export * from './map/mapUtils';
