+++ src/controllers/enrollmentReports/map/mapUtils/organization.ts
@@ -1,6 +1,6 @@
-import { Organization } from '../../../entity';
-import { EnrollmentReportRow } from '../../../template';
-import { BadRequestError } from '../../../middleware/error/errors';
+import { Organization } from '../../../../entity';
+import { EnrollmentReportRow } from '../../../../template';
+import { BadRequestError } from '../../../../middleware/error/errors';
 
 export const MISSING_PROVIDER_ERROR =
   'You uploaded a file with missing information.\nProvider name is required for every record in your upload. Make sure this column is not empty.';
