+++ src/routes/enrollmentReports.ts
@@ -14,6 +14,7 @@ import fs from 'fs';
 import { BatchUpload } from '../../client/src/shared/payloads';
 
 export const enrollmentReportsRouter = express.Router();
+const upload = multer({ dest: '/tmp/uploads' }).single('file');
 
 /**
  * /enrollment-reports/check POST
@@ -28,36 +29,29 @@ export const enrollmentReportsRouter = express.Router();
  *  - sends back an Error Dictionary mapping columns to how many errors
  *    of that kind occur in the file
  */
-const temp = multer({ dest: '/tmp/uploads' }).single('file');
 enrollmentReportsRouter.post(
   '/check',
-  temp,
+  upload,
   passAsyncError(async (req, res) => {
     return getManager().transaction(async (tManager) => {
       try {
         const reportRows = controller.parseUploadedTemplate(req.file);
-        const reportChildren: Child[] = await controller.mapRows(
+        const reportChildren = await controller.mapRows(
           tManager,
           reportRows,
           req.user,
           { save: false }
         );
 
-        // Need this line to create entities as the DB would see them.
-        // Since only given properties are copied into the entities,
-        // anything that winds up with missing info will set off
-        // validation errors, which we need to find nested errors in
-        // e.g. family address, enrollments, income dets, etc.
-        const dbChildren = tManager.create(Child, reportChildren);
         const childrenWithErrors = await Promise.all(
-          dbChildren.map(async (child) => {
+          reportChildren.map(async (child) => {
             return {
               ...child,
               validationErrors: await validate(child),
-              cascadeDeleteEnrollments: null,
-            };
+            } as Child;
           })
         );
+
         const errorDict = await controller.checkErrorsInChildren(
           childrenWithErrors
         );
@@ -94,7 +88,6 @@ enrollmentReportsRouter.post(
  * 		and saves that to the DB as an EnrollmentReport
  * 	- returns new EnrollmentReport on success
  */
-const upload = multer({ dest: '/tmp/uploads' }).single('file');
 enrollmentReportsRouter.post(
   '/',
   upload,
@@ -127,6 +120,7 @@ enrollmentReportsRouter.post(
           req.user,
           { save: true }
         );
+
         const report = await tManager.save(
           tManager.create(EnrollmentReport, { children: reportChildren })
         );
